make lib/todo_tutorial_web/controllers/user_controller.ex
